- Directory restructuration - Fix route, library - @todo need to accomplish the rest page

+++ resources/views/admin/partials/layout/form/delete.blade.php
@@ -1,8 +1,7 @@
 <div class="text-right">
-    @if($isUpdate === true and $enableDelete === true)
+    @if(isset($isUpdate) === true and  $isUpdate === true and $enableDelete === true)
         {!! Form::open(['method' => 'DELETE', 'url' => action($controllerName . '@destroy', $referenceValue)]) !!}
         <button onclick="return confirm('Are you sure to delete this record ?');" type="submit" class="btn btn-danger"><i class="fa fa-trash"></i> Delete This Record (Soft Delete)</button>
         {!! Form::close() !!}
     @endif
 </div>
-
